+++ src/main/java/com/kakao/cafe/web/controller/UserController.java
@@ -18,7 +18,6 @@ public class UserController {
 
 	private final UserService userService;
 
-	@Autowired
 	public UserController(UserService userService) {
 		this.userService = userService;
 	}
